requirements.txt
Deprecated==1.2.13
PyYAML==6.0
anyio==3.5.0
asgiref==3.5.0
backoff==1.11.1
beautifulsoup4==4.10.0
cacheout==0.13.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
decorator==5.1.1
fastapi==0.75.0
felix-flow==0.2.1
felix-library==0.2.3
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-appengine-logging==1.1.1
google-cloud-audit-log==0.2.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-logging==3.0.0
google-cloud-monitoring==2.12.0
google-cloud-pubsub==2.11.0
google-cloud-secret-manager==2.9.2
google-cloud-trace==1.8.0
google-cloud==0.34.0
google==3.0.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio-status==1.44.0
grpcio==1.44.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
monotonic==1.6
oauth2client
opentelemetry-api==1.13.0
opentelemetry-exporter-gcp-monitoring==1.4.0a0
opentelemetry-exporter-gcp-trace==1.4.0
opentelemetry-instrumentation-asgi==0.34b0
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-instrumentation-logging==0.34b0
opentelemetry-instrumentation-requests==0.34b0
opentelemetry-instrumentation==0.34b0
opentelemetry-propagator-gcp==1.4.0
opentelemetry-resourcedetector-gcp==1.4.0a0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
opentelemetry-util-http==0.34b0
packaging==21.3
posthog==1.4.9
proto-plus==1.22.1
protobuf==3.20.3
py==1.11.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
readerwriterlock==1.0.9
redis==4.3.4
requests==2.27.1
retry==0.9.2
rsa==4.8
segment-analytics-python==2.2.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
types-protobuf
types-redis
types-requests
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.1
websockets==10.2
wrapt==1.14.1
